requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
apscheduler>=3.10.4
anthropic>=0.28.0
chromadb>=0.5.0
playwright>=1.44.0
feedparser>=6.0.11
requests>=2.31.0
python-dotenv>=1.0.1
httpx>=0.27.0
beautifulsoup4>=4.12.0
numpy>=1.26.0
supabase>=2.4.0
